+++ e2e/console-debugging/jest.config.js
@@ -1,0 +1,6 @@
+require('./stdout-spy');
+
+module.exports = {
+  testEnvironment: 'node',
+  verbose: true,
+};